+++ src/main/java/seedu/rc4hdb/logic/parser/commandparsers/CommandParser.java
@@ -1,0 +1,17 @@
+package seedu.rc4hdb.logic.parser.commandparsers;
+
+import seedu.rc4hdb.logic.commands.Command;
+import seedu.rc4hdb.logic.parser.Parser;
+import seedu.rc4hdb.logic.parser.exceptions.ParseException;
+
+/**
+ * Represents a Parser that is able to parse user input into a {@code Command} of type {@code T}.
+ */
+public interface CommandParser<T extends Command> extends Parser<T> {
+
+    /**
+     * Parses {@code userInput} into a command and returns it.
+     * @throws ParseException if {@code userInput} does not conform the expected format
+     */
+    T parse(String userInput) throws ParseException;
+}
