+++ src/main/java/seedu/rc4hdb/logic/parser/commandparsers/EditCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.rc4hdb.logic.parser;
+package seedu.rc4hdb.logic.parser.commandparsers;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.rc4hdb.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
@@ -16,13 +16,16 @@ import java.util.Set;
 import seedu.rc4hdb.commons.core.index.Index;
 import seedu.rc4hdb.logic.commands.EditCommand;
 import seedu.rc4hdb.logic.commands.EditCommand.EditPersonDescriptor;
+import seedu.rc4hdb.logic.parser.ArgumentMultimap;
+import seedu.rc4hdb.logic.parser.ArgumentTokenizer;
+import seedu.rc4hdb.logic.parser.ParserUtil;
 import seedu.rc4hdb.logic.parser.exceptions.ParseException;
 import seedu.rc4hdb.model.tag.Tag;
 
 /**
  * Parses input arguments and creates a new EditCommand object
  */
-public class EditCommandParser implements Parser<EditCommand> {
+public class EditCommandParser implements CommandParser<EditCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the EditCommand
