+++ src/test/java/seedu/rc4hdb/logic/parser/CommandParserTestUtil.java
@@ -3,6 +3,7 @@ package seedu.rc4hdb.logic.parser;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import seedu.rc4hdb.logic.commands.Command;
+import seedu.rc4hdb.logic.parser.commandparsers.CommandParser;
 import seedu.rc4hdb.logic.parser.exceptions.ParseException;
 
 /**
@@ -14,8 +15,8 @@ public class CommandParserTestUtil {
      * Asserts that the parsing of {@code userInput} by {@code parser} is successful and the command created
      * equals to {@code expectedCommand}.
      */
-    public static void assertParseSuccess(Parser<? extends Command> parser, String userInput,
-            Command expectedCommand) {
+    public static void assertParseSuccess(CommandParser<? extends Command> parser, String userInput,
+                                          Command expectedCommand) {
         try {
             Command command = parser.parse(userInput);
             assertEquals(expectedCommand, command);
@@ -28,7 +29,7 @@ public class CommandParserTestUtil {
      * Asserts that the parsing of {@code userInput} by {@code parser} is unsuccessful and the error message
      * equals to {@code expectedMessage}.
      */
-    public static void assertParseFailure(Parser<? extends Command> parser, String userInput, String expectedMessage) {
+    public static void assertParseFailure(CommandParser<? extends Command> parser, String userInput, String expectedMessage) {
         try {
             parser.parse(userInput);
             throw new AssertionError("The expected ParseException was not thrown.");
