+++ src/test/java/seedu/rc4hdb/logic/parser/DeleteCommandParserTest.java
@@ -8,6 +8,7 @@ import static seedu.rc4hdb.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.jupiter.api.Test;
 
 import seedu.rc4hdb.logic.commands.DeleteCommand;
+import seedu.rc4hdb.logic.parser.commandparsers.DeleteCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
